backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
snowflake-connector-python==3.12.4
cryptography==44.0.2
google-generativeai==0.8.3
google-genai>=0.8.0
google-cloud-documentai==2.33.0
elevenlabs==1.51.0
python-dotenv==1.0.1
python-docx==1.1.2
reportlab==4.2.5
pymongo[srv]==4.10.1
openpyxl==3.1.5
httpx==0.28.1
aiohttp==3.11.11
xmltodict==0.14.2
